% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{marathonR-package}
\alias{marathonR}
\alias{marathonR-package}
\title{Documentation for the \pkg{marathonR} package}
\description{
This packages provides methods for the uniform sampling of binary matrices with restricted row and column sums.
For this purpose, the package provides two functions for the uniform sampling of 
  a) binary matrices with prescribed row and column sums.
  b) binary matrices whose row and column sums are restricted by lower and upper bounds.
}
